skip already downloaded files

+++ fetch-data/download_repos.py
@@ -16,13 +16,16 @@ def download_repo_as_zip(repo_name, branch_url, save_path):
         if response.status_code == 200:
             # Define the zip file path where the repo will be saved
             zip_file_path = os.path.join(save_path, f"{repo_name}.zip")
-            
+
+            if not os.path.exists(zip_file_path):            
             # Write the content of the request to a file
-            with open(zip_file_path, 'wb') as zip_file:
-                for chunk in response.iter_content(chunk_size=128):
-                    zip_file.write(chunk)
-            
-            print(f"Successfully downloaded {repo_name} as {zip_file_path}")
+                with open(zip_file_path, 'wb') as zip_file:
+                    for chunk in response.iter_content(chunk_size=128):
+                        zip_file.write(chunk)
+                
+                print(f"Successfully downloaded {repo_name} as {zip_file_path}")
+            else:
+                print(f"Skipping {repo_name} as {zip_file_path} already exists")
         else:
             print(f"Failed to download {repo_name}. HTTP Status Code: {response.status_code}")
     
@@ -49,7 +52,7 @@ def download_repositories(csv_file, save_path):
 
 if __name__ == "__main__":
     # Replace 'repos.csv' with the actual path to your CSV file
-    csv_file_path = './fetch-data/repos.csv'
+    csv_file_path = 'repos.csv'
     
     # Replace with the desired path where you want to save the ZIP files
     save_directory = 'downloaded_repos'
